test(unit): rename
